+++ src/keypin/core.clj
@@ -8,6 +8,7 @@
 
 
 (ns keypin.core
+  "Public API for the core functionality."
   (:require
     [clojure.edn     :as edn]
     [clojure.string  :as string]
@@ -15,7 +16,7 @@
     [keypin.type     :as t]
     [keypin.util     :as u])
   (:import
-    [clojure.lang IDeref ILookup]
+    [clojure.lang Associative IDeref]
     [java.io OutputStream Writer]
     [java.util Map Properties]
     [keypin Config ConfigIO Logger Mapper PropertyConfigIO]))
@@ -25,11 +26,11 @@
 
 
 (defn lookup-key
-  "Look up a key in a map or something that implements clojure.lang.ILookup."
+  "Look up a key in an associative data structure - a map or something that implements `clojure.lang.Associative`."
   [the-map the-key validator description value-parser default-value? default-value not-found]
   (when-not (or (instance? Map the-map)
-              (instance? ILookup the-map))
-    (i/illegal-arg (format "Key %s looked up in a non map (or clojure.lang.ILookup) object: %s"
+              (instance? Associative the-map))
+    (i/illegal-arg (format "Key %s looked up in a non map (or clojure.lang.Associative) object: %s"
                      (pr-str the-key) the-map)))
   (let [value (if (contains? the-map the-key)
                 (->> (get the-map the-key)
@@ -43,13 +44,13 @@
 
 
 (defn lookup-keypath
-  "Look up a key path in a map or something that implements clojure.lang.ILookup."
+  "Look up a key path in an associative data structure - a map or something that implements `clojure.lang.Associative`."
   [the-map ks validator description value-parser default-value? default-value not-found]
   (let [value (loop [data the-map
                      path ks]
                 (when-not (or (instance? Map data)
-                            (instance? ILookup data))
-                  (i/illegal-arg (format "Key path %s looked up in a non map (or clojure.lang.ILookup) object: %s"
+                            (instance? Associative data))
+                  (i/illegal-arg (format "Key path %s looked up in a non map (or clojure.lang.Associative) object: %s"
                                    (pr-str path) (pr-str data))))
                 (let [k (first path)]
                   (if-not (contains? data k)
@@ -105,9 +106,13 @@
 
 (defn realize-config
   "Realize config by applying variable substitution, if any.
-  Options:
-  :logger        (object) instance of keypin.Logger, defaults to printing to *err*
-  :config-mapper (object) instance of keypin.Mapper, default: keypin.Mapper/DEFAULT"
+
+  ### Options
+
+  | Kwarg          | Type   | Description                 | Default                 |
+  |----------------|--------|-----------------------------|-------------------------|
+  |`:logger`       | object | instance of `keypin.Logger` | prints to `*err*`       |
+  |`:config-mapper`| object | instance of `keypin.Mapper` | `keypin.Mapper/DEFAULT` |"
   ([config]
     (realize-config config {}))
   ([config {:keys [logger config-mapper]
@@ -118,13 +123,17 @@
 
 
 (defn read-config
-  "Read config file(s) returning a java.util.Map instance.
-  Options:
-  :parent-key     (string)   key to identify the parent filenames having K/V pairs, default: \"parent.filenames\"
-  :logger         (object)   instance of keypin.Logger, defaults to printing to *err*
-  :config-readers (list/vec) collection of keypin.ConfigIO instances, default: for Properties and EDN files
-  :media-readers  (list/vec) collection of keypin.MediaReader instances, default: for Filesystem and Classpath
-  :realize?       (boolean)  whether realize the template variables in the string, default: true"
+  "Read config file(s) returning a `java.util.Map` instance.
+
+  ### Options
+
+  | Kwarg           | Type     | Description                                  | Default                         |
+  |-----------------|----------|----------------------------------------------|---------------------------------|
+  |`:parent-key`    | string   | key to identify the parent config filenames  | `\"parent.filenames\"`          |
+  |`:logger`        | object   | instance of `keypin.Logger`                  | prints to `*err*`               |
+  |`:config-readers`| list/vec | collection of `keypin.ConfigIO` instances    | for `.properties`, `.edn` files |
+  |`:media-readers` | list/vec | collection of `keypin.MediaReader` instances | for filesystem and classpath    |
+  |`:realize?`      | boolean  | whether realize template variables in string | `true`                          |"
   (^Map [config-filenames]
     (read-config config-filenames {}))
   (^Map [config-filenames {:keys [^String parent-key logger config-readers media-readers realize?]
@@ -147,10 +156,14 @@
 
 (defn write-config
   "Write config to a specified file.
-  Options:
-  :logger         (object)   instance of keypin.Logger, defaults to printing to *err*
-  :config-writers (list/vec) collection of keypin.ConfigIO instances, default: for Properties and EDN files
-  :escape?        (boolean)  whether escape values when writing, default: true"
+
+  ### Options
+
+  | Kwarg           | Type     | Description                               | Default                         |
+  |-----------------|----------|-------------------------------------------|---------------------------------|
+  |`:logger`        | object   | instance of `keypin.Logger`               | prints to `*err*`               |
+  |`:config-writers`| list/vec | collection of `keypin.ConfigIO` instances | for `.properties`, `.edn` files |
+  |`:escape?`       | boolean  | whether escape values when writing        | `true`                          |"
   ([config-filename config]
     (write-config config-filename config {}))
   ([config-filename config {:keys [logger config-writers escape?]
@@ -165,59 +178,131 @@
 
 
 (defn make-key
-  "Create a key that can be looked up in a java.util.Map/Properties or clojure.lang.ILookup (map, vector) instance. The
-  following optional keys are supported:
-  :lookup  - The function to look the key up:
-             (fn [the-map the-key validator description value-parser default-value? default-value
-                  (fn not-found-fn [not-found-message])])
-             default: ordinary key look up
-  :parser  - The value parser function (args: key, value)
-  :default - Default value to return if key is not found
-  :sysprop - System property name that can override the config value (before parsing)
-  :envvar  - Environment variable that can override the config value and system property (before parsing)
-  :source  - Source or container (of reference type, e.g. atom/promise etc.) of key/value pairs"
-  [the-key validator description {:keys [lookup parser default sysprop envvar source]
-                                  :or {lookup lookup-key
-                                       parser i/identity-parser}
-                                  :as options}]
-  (t/->KeyAttributes
-    the-key validator description parser
-    (if   (contains? options :default) true false)
-    (when (contains? options :default) default)
-    (fn [the-map the-key validator description value-parser default-value? default-value]
-      (cond
-        (and envvar  (System/getenv envvar))       (value-parser the-key (System/getenv envvar))
-        (and sysprop (System/getProperty sysprop)) (value-parser the-key (System/getProperty sysprop))
-        :otherwise (lookup the-map the-key validator description value-parser default-value? default-value
-                     (if (or envvar sysprop)
-                       (fn [message] (str
-                                       (when-not envvar  (format "Environment variable '%s' is not defined. " envvar))
-                                       (when-not sysprop (format "System property '%s' is not defined. " sysprop))
-                                       message))
-                       identity))))
-    (if (nil? source)
-      nil
-      (do
-        (i/expected #(instance? IDeref %) "a key/value-source of type clojure.lang.IDeref (atom/promise etc.)" source)
-        source))))
+  "Create a key that can be looked up in a config store (`keypin.type/IStore`, `java.util.Map`, `java.util.Properties`
+  or map/vector) instance.
+
+  ### Options
+
+  | Kwarg    | Description                 |
+  |----------|-----------------------------|
+  |`:the-key`|(Required) Key to be used for looking up value                                |
+  |`:pred`   |Validator function `(fn [value]) -> boolean` - default fn just returns true   |
+  |`:desc`   |Description string for the config - default: `\"No description\"`             |
+  |`:lookup` |Function to look the key up (details below) - default: [[lookup-key]] - ordinary key look up |
+  |`:parser` |The value parser fn `(fn [key raw-value]) -> parsed-value`, e.g. [[keypin.util/str->long]]   |
+  |`:default`|Default value when key is not found (subject to validator, unspecified => no default value)  |
+  |`:sysprop`|System property name that can override the config value (before parsing)                     |
+  |`:envvar` |Environment variable that can override the config value and system property (before parsing) |
+  |`:source` |Source or container (of reference type, e.g. atom/agent/promise etc.) of key/value pairs     |
+
+  ### Lookup function
+
+  ```
+  (fn [store
+       key
+       validator
+       description
+       value-parser
+       default-value?
+       default-value
+       not-found-handler]) -> value
+
+  ;; validator is a predicate function: (fn [parsed-value]) -> boolean
+  ;; not-found-handler is (fn [not-found-message]), called when the store does not have the key
+  ```
+
+  See: [[lookup-key]], [[lookup-keypath]]"
+  [{:keys [the-key  ; required
+           pred     ; validator
+           desc     ; description
+           lookup
+           parser
+           default
+           sysprop
+           envvar
+           source]
+    :or {pred   i/return-true
+         desc   "No description"
+         lookup lookup-key
+         parser i/identity-parser}
+    :as options}]
+  (let [validator   pred
+        description desc]
+    (i/expected some?   "Non-nil key for lookup (option :the-key)" the-key)
+    (i/expected fn?     "Validator function under (option :pred)"  validator)
+    (i/expected string? "Config description string (option :desc)" description)
+    (i/expected fn?     "Lookup function (option :lookup)"         lookup)
+    (i/expected fn?     "Parser function (option :parser)"         parser)
+    (t/->KeyAttributes
+      the-key validator description parser
+      (if   (contains? options :default) true false)
+      (when (contains? options :default) default)
+      (fn [the-map the-key validator description value-parser default-value? default-value]
+        (cond
+          (and envvar  (System/getenv envvar))       (value-parser the-key (System/getenv envvar))
+          (and sysprop (System/getProperty sysprop)) (value-parser the-key (System/getProperty sysprop))
+          :otherwise (lookup the-map the-key validator description value-parser default-value? default-value
+                       (if (or envvar sysprop)
+                         (fn [message] (str
+                                         (when-not envvar  (format "Environment variable '%s' is not defined. " envvar))
+                                         (when-not sysprop (format "System property '%s' is not defined. " sysprop))
+                                         message))
+                         identity))))
+      (if (nil? source)
+        nil
+        (do
+          (i/expected #(instance? IDeref %) "a key/value-source of type clojure.lang.IDeref (atom/promise etc.)" source)
+          source)))))
 
 
 (defmacro defkey
-  "Define one or more keys as vars using argument vectors. Every argument vector must have one of the following arities:
+  "Define one or more keys as `defn` vars using argument vectors, which could be invoked as `(keydef store)` to
+  fetch the value of the key. Every argument vector must have one of the following arities:
+
+  ```
   [key]
   [key options]
   [key validator description]
   [key validator description options]
-  See `make-key` for details. First argument to `defkey` can optionally be a base option-map for all argument vectors.
-  Examples:
+  ```
+
+  ### Options
+
+  | Kwarg       | Description |
+  |-------------|-------------|
+  |`:pred`      |Validator function `(fn [value]) -> boolean` - default fn just returns `true`  |
+  |`:desc`      |Description string for the config - default: `\"No description\"`              |
+  |`:lookup`    |Function to look the key up (details below) - default: [[lookup-key]] - ordinary key look up |
+  |`:parser`    |The value parser fn `(fn [key raw-value]) -> parsed-value`, e.g. [[keypin.util/str->long]]   |
+  |`:default`   |Default value when key is not found (subject to validator, unspecified => no default value)  |
+  |`:sysprop`   |System property name that can override the config value (before parsing)                     |
+  |`:envvar`    |Environment variable that can override the config value and system property (before parsing) |
+  |`:source`    |Source or container (of reference type, e.g. atom/agent/promise etc.) of key/value pairs     |
+  |`:pre-xform` |Middleware function `(fn [option-map]) -> option-map` used before key definition is created  |
+  |`:post-xform`|Middleware function `(fn [keypin.type.KeyAttributes]) -> keypin.type.KeyAttributes`          |
+
+  The `validator` is a predicate `(fn [parsed-value]) -> boolean` that returns `true` for valid values,
+  `false` otherwise.
+
+  First argument to `defkey` can optionally be a base option-map for all argument vectors.
+
+  ### Examples
+
+  ```
   (defkey
     ip   [:ip-address]
-    port [:port #(< 1023 % 65535) \"Server port\" {:parser str->int :default 3000}])
+    port [:port #(< 1023 % 65535) \"Server port\" {:parser keypin.util/str->int :default 3000}])
+
+  (port {:port \"3009\"})  ; returns 3009
+  (port {})              ; returns 3000
+
   (defkey
     {:lookup lookup-key}
     ip   [\"server.ip.address\"]
-    port [\"server.port\" #(< 1023 % 65535) \"Server port\" {:parser str->int :default 3000}])
-  See: make-key"
+    port [\"server.port\" #(< 1023 % 65535) \"Server port\" {:parser keypin.util/str->int :default 3000}])
+  ```
+
+  See: [[make-key]]"
   [the-sym arg-vec & more]
   (let [options (if (odd? (count more))
                   (i/expect-arg the-sym map? ["Expected an option map, found" (pr-str the-sym)])
@@ -230,9 +315,9 @@
                        (i/expect-arg each-sym symbol? ["Expected a symbol to define var, but found" (pr-str each-sym)])
                        (i/expect-arg each-vec vector? ["Expected a vector to create key, but found" (pr-str each-vec)])
                        [each-sym (case (count each-vec)
-                                   1 (conj each-vec (constantly true) "No description" options)
+                                   1 (conj each-vec i/return-true "No description" options)
                                    2 (let [{:keys [pred desc]
-                                            :or {pred (constantly true)
+                                            :or {pred i/return-true
                                                  desc "No description"}
                                             :as spec-opts} (let [[each-key each-opts] each-vec]
                                                              (i/expect-arg each-opts map?
@@ -249,15 +334,21 @@
                                                                (merge options each-opts)))
                                    (i/illegal-arg ["Expected 1, 2, 3 or 4 elements as arguments for defkey, but found"
                                                    (pr-str each-vec)]))]))
-                (map (fn [[each-sym each-vec]]
-                       (let [descrip (nth each-vec 2 "No description")
-                             options (nth each-vec 3 {})  ; 4th element is option map
+                (map (fn [[each-sym [the-key validator descrip options :as each-vec]]]
+                       (let [;; descrip (nth each-vec 2 "No description")
+                             ;; options (nth each-vec 3 {})  ; 4th element is option map
                              arities (if (:source options)
                                        ''([] [config-map] [config-map not-found])
                                        ''([config-map] [config-map not-found]))
                              meta-fn (partial merge {:arglists arities :doc descrip})
-                             def-sym (vary-meta each-sym meta-fn)]
-                         `(def ~def-sym (make-key ~@each-vec))))))]
+                             def-sym (vary-meta each-sym meta-fn)
+                             arg-map (conj (get each-vec 3 {})
+                                       {:the-key (get each-vec 0)
+                                        :pred    (get each-vec 1)
+                                        :desc    (get each-vec 2)})
+                             pre-xform (:pre-xform options identity)
+                             post-xform (:post-xform options identity)]
+                         `(def ~def-sym (~post-xform (make-key (~pre-xform ~arg-map))))))))]
     `(do ~@pairs)))
 
 
@@ -265,14 +356,18 @@
 
 
 (defmacro letval
-  "Like let, except in which the left hand side is a destructuring map, right hand side is the argument to key finder.
-  Beside symbols, the destructuring map optionally supports :defs (symbols bound to key finders) and :as keys.
-  Example:
+  "Like `let`, except in which the left hand side is a destructuring map, right hand side is the argument to key finder.
+  Beside symbols, the destructuring map optionally supports `:defs` (symbols bound to key finders) and `:as` keys.
+
+  ### Example
+
+  ```
   (letval [{:defs [foo bar] ; foo, bar are key finders
             baz baz-key     ; baz-key is a key finder
             :as m} {:foo 10 :bar 20 :baz 30}]
     ;; foo, bar and baz are now bound to values looked up in the map
-    ...)"
+    ...)
+  ```"
   [bindings & body]
   (i/expect-arg bindings vector? ["Expected a binding vector, but found" (pr-str bindings)])
   (i/expect-arg (count bindings) even? ["Expected even number of binding forms, but found" (pr-str bindings)])