+++ src/keypin/internal.clj
@@ -8,6 +8,9 @@
 
 
 (ns keypin.internal
+  "This is an internal namespace. Code here is subject to change.
+
+  Utility functions."
   (:require
     [clojure.string :as string])
   (:import
@@ -62,3 +65,24 @@
   [^long n k]
   (throw (ArityException. n (str "key definition " (pr-str k)
                               ", allowed: [] [the-map] [the-map default-value]"))))
+
+
+(defn now-millis
+  (^long []
+    (System/currentTimeMillis))
+  (^long [^long since]
+    (unchecked-subtract (now-millis) since)))
+
+
+(defn as-str
+  [x]
+  (if (instance? clojure.lang.Named x)
+    (if-let [xns (namespace x)]
+      (str xns "/" (name x))
+      (name x))
+    (str x)))
+
+
+(defn return-true
+  [& _]
+  true)